Drag System
To be placed on icons representing the item in a slot. Allows the item to be dragged into other slots.
Create a subclass for the type you want to be draggable. Then place on the UI element you want to make draggable.

After the item is dropped it will be automatically return to the original UI parent. It is the job of components implementing IDragableContainer, IDragDestination and IDrag to update the interface after a drag